Unoapp/requirement.txt
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.0.7
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.3
idna==2.8
lxml==4.3.2
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
virtualenv==16.4.3
webencodings==0.5.1
